FeedForwardNN/weather_forecasting/requirements.txt
tensorflow-macos>=2.12.0
tensorflow-metal>=0.7.0
numpy>=1.22.0,<1.24.0
pandas>=2.0.0
streamlit>=1.31.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
plotly>=5.18.0
statsmodels>=0.14.0
